CI nuevo arreglo 24 // Eliminación de test auth.test.js y staff.test.js debido a interposición de codigo entre si; Creación de archivo api.test.js para consolidación de archivos test borrados.
